Add test for bean show page and organize
Adds a feature test for a guest visiting the bean show page.  Adds show method for
to beans_controller and route for same.  Adds show view for beans; it was patterned
after the root page to keep theme contiguous.  Made a beans folder in spec/features
to keep in line with organization.  Moved item_index_spec into same; bean show test
is stored in same.