Add scenario to bean show page feature test
Added a scenario to the bean show page feature test.  Guest will be
able to click 'Add Item' on show page and have the item added to
their cart.  Copied the add item button from the same on the
bean index page; changed it to reference '@bean.id' so that the
correct id would be passed in params.  Test is currently passing.