Add Orders
Adds orders table with attributes as contents, total_price, and a reference to user
. One-to-many relationship exists between user and orders.  Appropriate methods
added to user and order models for same.  Model testing added to user to check
validation of contents presence as well as relationship to user. Feature testing
confirms that an unlogged in user can add items to cart but must login or create
an account in order to checkout.  Feature testing confirms that at checkout
an order is added to the database that references the ordering user.