Adds ability for user to view all past orders
Add index route to orders in routes.rb.  Add index method to orders_controller.
Index method queries the orders of the current_user to prevent cross-user
order-viewing.  Add index.html.erb to orders views.  Same uses a striped-table
to display the order id, order item-count, order datetime, the total price of
the order, and a button which links to the show page of the individual order.
All tests are currently passing and local testing shows the view to be
rendering with anticipated formatting.