Add feature user may view a single order in history
Add feature test to prove user can view a single order in their history.
Restyled orders show page to conform to user story provided. Add method
to bean to determine its subtotal with a given quantity.  View includes
simple ruby logic to determine if a condition applies to render the
orders update to status time. Same utilizes a new method in the order
model to determine if it's status is completed or cancelled. All
feature tests are presently passing. It was necessary to modify
user checks out feature test to conform to new styling of
order show page.