Add admin viewing single order test
Adds basic feature test for admin viewing a single order.
Adds route to admin users controller for same.
Adds show page which is formatted with Bootstrap to
display appropriate information. After this commit
going to add name, street, city, state, and zip
to user attributes. This is necessary to conform to
this user story.