Add dragonfly for images
Implements dragonfly gem for the purpose of
attaching images to beans. Installed gem,
created migration for same, modified seed
file for bean generation, modified image tags
in beans index view. All appear to be functioning
appropriately; however, the images/boxes need to
be formatted for uniformity, and all tests
involving bean generation will need to be
modified as the generation of a bean now
requires the presence of an image.