updates admin dashboard
Updates admin dashboard to accept order beans in the
dashboard view. Fixed status change buttons to only
show in the proper scenarios according to the spec.
Adds logic to the navbar to show a link to the regular
dashboard if a user and the admin dashboard if user is
an admin.